refactor: move decrementItemFromInventory to its own file
