+++ src/game-logic/reducers/common.js
@@ -1,34 +1,6 @@
 import { inventorySpaceRemaining } from '../../utils'
 
 /**
- * @param {farmhand.state} state
- * @param {string} itemId
- * @param {number} [howMany=1]
- * @returns {farmhand.state}
- */
-export const decrementItemFromInventory = (state, itemId, howMany = 1) => {
-  const inventory = [...state.inventory]
-  const itemInventoryIndex = inventory.findIndex(({ id }) => id === itemId)
-
-  if (itemInventoryIndex === -1) {
-    return state
-  }
-
-  const { quantity } = inventory[itemInventoryIndex]
-
-  if (quantity > howMany) {
-    inventory[itemInventoryIndex] = {
-      ...inventory[itemInventoryIndex],
-      quantity: quantity - howMany,
-    }
-  } else {
-    inventory.splice(itemInventoryIndex, 1)
-  }
-
-  return { ...state, inventory }
-}
-
-/**
  * Only adds as many items as there is room in the inventory for unless
  * allowInventoryOverage is true.
  * @param {farmhand.state} state
